+++ Assets/@Project/_Prototypes/OutpostAISystem/Scripts/DamageableOutpost.cs
@@ -10,7 +10,7 @@ namespace DuneRiders.OutpostAI {
 	{
 		[SerializeField] StructureHealth healthState;
 
-		public override void Damage(int healthPoints)
+		public override void Damage(float healthPoints)
 		{
 			healthState.health -= healthPoints;
 		}
